app/requirements.txt
mt5linux
python-dotenv
clickhouse-driver
esmerald 
requests
retry==0.9.2
